Extract validation methods to their own class
